+++ demo3PNG/src/types/index.ts
@@ -1,4 +1,4 @@
 import { layers } from "../configs";
 
 export type LayerName = (typeof layers)[number];
-export type Attributes = Record<LayerName, string>;
+export type Attributes = Record<string, string>;
